requirements.txt
# Core Machine Learning & Vision
torchvision==0.19.0
opencv-contrib-python==4.10.0.84
albumentations==1.4.4

# Data Processing & Visualization
matplotlib==3.8.2
pycocotools==2.0.7
numpy<2.0.0
scipy>=1.5.4
scikit-learn==1.3.2
scikit-image==0.23.2
av>=8.0.3
pandas

# Utilities
pyyaml==6.0.1
pydantic==2.6.4
tqdm
smplx